last

+++ ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: oumimoun <oumimoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 15:49:18 by oumimoun          #+#    #+#             */
-/*   Updated: 2023/11/07 11:26:56 by oumimoun         ###   ########.fr       */
+/*   Updated: 2023/11/15 09:55:11 by oumimoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
 	unsigned int	i;
 
@@ -26,5 +26,5 @@ unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
 		}
 		dest[i] = '\0';
 	}
-	return ((unsigned int)ft_strlen(src));
+	return (ft_strlen(src));
 }
